Fixes in the tool to support the airbase dataset

+++ src/main/java/dk/aau/cs/qweb/pec/data/JenaTDBDatabaseConnection.java
@@ -80,8 +80,8 @@ public class JenaTDBDatabaseConnection implements RDFCubeDataSource {
 			if (Config.getBudgetPercentages().size() == 1) {
 				long budgetInPercentage = Config.getBudgetPercentages().get(0);
 				long dbSizeInKB = FileUtils.sizeOfDirectory(new File(dbLocation)) / 1024;
-				long node2NodeIdCacheSize = (dbSizeInKB / 100) * budgetInPercentage * (1 / 6);
-				long nodeId2NodeCacheSize = (dbSizeInKB / 100) * budgetInPercentage * (5 / 6);
+				long node2NodeIdCacheSize = (long)((dbSizeInKB / 100) * budgetInPercentage * (1.0 / 6));
+				long nodeId2NodeCacheSize = (long)((dbSizeInKB / 100) * budgetInPercentage * (5.0 / 6));
 				params = StoreParams.builder()
 						.blockReadCacheSize(0)
 						.blockWriteCacheSize(0)
